Updated interface.py functionality
Now displays 'Set Temperature' value and the difference between the current temperature and this value. The current temperature is now only displayed once the 'Get Temperature' button has been pressed. The 'Toggle Heating' and 'Toggle Cooling' displays are now updated to show the current state that the heating/cooling is in respectively. A small piece of text has been added which is displayed when the user presses 'How to use'. This informs the user on how to use the application.