Update interface.py
Now includes set time option like in admin control panel. This option is only presented to the user once they have set a specific temperature. A small timer is added next to the time entry box to display how much time is left. The time value is currently input in seconds for convenience , Later this may be input as number of hours to follow a more normal heating system approach and converted to seconds behind the scene.