Created new_temperature_physics
Created a new function to replace old temperature_physics, new function is run when called, it calculates using a timer, _start_time and _time_taken.
_time_taken = "Time right now" - _start_time.

When the function runs through it no longer waits _tminus_temp_increase or equivalent, blocking, but now calculated the time taken and if _time_taken % _tminus_temp_increase = 0 or equivalent it will add to _zone_temp _time_taken / variable _tminus_temp_increase

The state checks and function is much the same as old function except this function only has to be run when an update in _zone_temp is required instead of running constantly and blocking.